Introduces the concept of a Pipeline.
A Pipeline represents the entire construction of an aggregator, senders,
endpoints, etc. Its interface is just that of the original Aggregator,
so Aggregator now implements Pipeline.

Additionally, PipelineComponent is any component with a Close method that
satisfies a documented protocol: must shut itself down first and wait, then
shutdown all adjacent components and wait. A Pipeline is itself a
PipelineComponent.

This refactoring replaces the original App with a pipeline builder. Instead
of returning an App object, the builder returns a Pipeline (essentially an
Aggregator) and at shutdown, Pipeline.Close() is used instead of
App.Shutdown().

Change-Id: Ie39ee8d8aa972ffdd4a16204cb7b7dbe019d9fe5