Removes the Sender Prepare/Send split, replacing it with only Send.
The original idea behind Prepare was to give a chance for Dispatcher to
abort sending to all endpoints if one endpoint generated an early
failure (for example, it failed to render an EndpointReport for some
reason).

This added unnecessary complexity. Instead, we should just try to send
to as many endpoints as possible while also registering failures when
not all are successful.

Sender now just has a single Send method which takes a single report. If
a call to Send results in 1 endpoint failing to render the report, but
two others succeeding, the report will be received by two endpoints but
the agent will register a failure.