Config changes to support flexible model (#19)
Config changes to support flexible model.

Metric configuration has expanded to include both a type configuration
and a collection of destination endpoints. For example, a metrics
section that used to look like this:

metrics:
  bufferSeconds: 10
  definitions:
  - name: int-metric
    type: int

might now look like this:

metrics:
- name: int-metric
  type: int
  endpoints:
  - name: disk
  - name: servicecontrol
  reported:
    bufferSeconds: 10

In the above example, `reported` specifies to the metric value source
type. A `reported` metric is one whose values are specified by an
external source, meaning the HTTP interface. Other types in the future
will include `heartbeat`, etc. Only one source type can be specified per
metric.

The `endpoints` section is a list of objects rather than a simple list
of string IDs. Today the only endpoint parameter is the name, but in the
future this enables specifying endpoint-specific metric options, such as
a different metric name to use when reporting to a particular endpoint.