Moves Heartbeat metric to a separate "source".
Sources are now more like autonomous processes that run within the
agent, rather than metric types. A Heartbeat source is configured with
the Metric name, interval and value. The referenced metric must also be
defined in the metrics section.

Additionally, there is now a way to define a passthrough, non-aggregated
metric.