Adds a heartbeat metric source (#23)
A heartbeat is a source that reports a pre-defined metric value at a configured interval. This can be used to, for example, report a recurring metric such as instance-seconds as long as the agent is running.

A source is like an autonomous processes that run within the agent. A heartbeat source is configured with a metric name, interval and value. The referenced metric must also be defined in the metrics section.

Additionally, there is now a way to define a passthrough, non-aggregated metric. Potentially useful (although not required) for heartbeats since a heartbeat has its own interval.

Example:

metrics:
- name: requests
  type: int
  aggregation:
    bufferSeconds: 20
  endpoints:
  - name: on_disk
- name: instance-seconds
  type: int
  passthrough: {}
  endpoints:
  - name: on_disk

endpoints:
- name: on_disk
  disk:
    reportDir: /tmp/disk
    expireSeconds: 3600

sources:
- name: instance-seconds
  heartbeat:
    metric: instance-seconds
    intervalSeconds: 10
    value:
      intValue: 10